refactor: 200OK json으로 응답변경
