+++ src/main/java/com/damyo/alpha/api/auth/exception/AuthException.java
@@ -1,9 +1,7 @@
 package com.damyo.alpha.api.auth.exception;
 
-import com.damyo.alpha.global.exception.error.BaseException;
-import com.damyo.alpha.global.exception.error.ErrorCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import com.damyo.alpha.global.response.exception.error.BaseException;
+import com.damyo.alpha.global.response.exception.error.ErrorCode;
 
 
 public class AuthException extends BaseException {
