+++ src/main/java/com/damyo/alpha/api/smokingarea/exception/AreaErrorCode.java
@@ -1,7 +1,7 @@
 package com.damyo.alpha.api.smokingarea.exception;
 
 
-import com.damyo.alpha.global.exception.error.ErrorCode;
+import com.damyo.alpha.global.response.exception.error.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
