+++ src/main/java/com/damyo/alpha/api/smokingarea/exception/AreaException.java
@@ -1,7 +1,7 @@
 package com.damyo.alpha.api.smokingarea.exception;
 
-import com.damyo.alpha.global.exception.error.BaseException;
-import com.damyo.alpha.global.exception.error.ErrorCode;
+import com.damyo.alpha.global.response.exception.error.BaseException;
+import com.damyo.alpha.global.response.exception.error.ErrorCode;
 
 public class AreaException extends BaseException {
     public AreaException(ErrorCode errorCode) {
