+++ src/main/java/com/damyo/alpha/api/star/exception/StarException.java
@@ -1,9 +1,7 @@
 package com.damyo.alpha.api.star.exception;
 
-import com.damyo.alpha.global.exception.error.BaseException;
-import com.damyo.alpha.global.exception.error.ErrorCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import com.damyo.alpha.global.response.exception.error.BaseException;
+import com.damyo.alpha.global.response.exception.error.ErrorCode;
 
 public class StarException extends BaseException {
