+++ src/main/java/com/damyo/alpha/api/user/exception/UserException.java
@@ -1,10 +1,8 @@
 package com.damyo.alpha.api.user.exception;
 
 
-import com.damyo.alpha.global.exception.error.BaseException;
-import com.damyo.alpha.global.exception.error.ErrorCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import com.damyo.alpha.global.response.exception.error.BaseException;
+import com.damyo.alpha.global.response.exception.error.ErrorCode;
 
 public class UserException extends BaseException {
     public UserException(ErrorCode errorCode) {
