+++ src/main/java/com/damyo/alpha/global/response/CommonSuccessApiResponse.java
@@ -1,0 +1,7 @@
+package com.damyo.alpha.global.response;
+
+public record CommonSuccessApiResponse(
+        String message
+) {
+
+}
