+++ src/main/java/com/damyo/alpha/global/response/exception/error/BaseException.java
@@ -1,4 +1,4 @@
-package com.damyo.alpha.global.exception.error;
+package com.damyo.alpha.global.response.exception.error;
 
 import lombok.Getter;
