+++ src/main/java/com/damyo/alpha/global/response/exception/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.damyo.alpha.global.exception.error;
+package com.damyo.alpha.global.response.exception.error;
 
 public record ErrorResponse(
         String code,
